refactor(liblili2json): move challenge formatters to challenge_types.py
